Fix ResourceManager race condition by using a callback approach
If two planes were both waiting for the same resource to be loaded,
they both kept the iterator to Future. As soon as the model was loaded,
the first plane got the result from the future and erased it from the
vector. This invalidated the iterator of the second plane and causes
in undefined behaviour.
Instead of polling the future state every frame, use a passive
callback approach. Once the worker is finished, a callback will be
called to transfer the handle.